requirements.txt
APScheduler==3.9.1
asgiref==3.3.1
astroid==2.11.5
backports.zoneinfo==0.2.1
boto3==1.23.1
botocore==1.26.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
dill==0.3.4
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-storages==1.12.3
django-webpack-loader==0.7.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
isort==5.10.1
jmespath==1.0.0
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
prettytable==3.3.0
psycopg2==2.8.6
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.1.0
wrapt==1.14.1
